+++ aqueduct/executor/spark/ingress.py
@@ -19,9 +19,10 @@ if TYPE_CHECKING:
 
 from aqueduct.parser.models import Module
 from aqueduct.executor.path_keys import PATHLESS_INGRESS_FORMATS
+from aqueduct.errors import AqueductError
 
 
-class IngressError(Exception):
+class IngressError(AqueductError):
     """Raised when an Ingress module fails to read."""